pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "hakowan"
authors = [{name = "Qingnan Zhou", email = "qnzhou@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dynamic = ["description", "version"]
dependencies=[
    "pytest",
    "mitsuba",
    "numpy>=1.21",
    "scipy>=1.8.0",
    "lagrange-open",
    ]
requires-python = ">=3.10"


[mypy]
plugins = "numpy.typing.mypy_plugin"
